requirements.txt
streamlit
opencv-python-headless
numpy
Pillow
pytesseract
PyPDF2
